fix: boardInit
